requirements.txt
apiai==1.2.3
attrs==18.1.0
Babel==2.5.3
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
facebook-sdk==2.0.0
Flask==0.12.2
Flask-Babel==0.11.2
Flask-JSGlue==0.3.1
Flask-Mail==0.9.1
Flask-WTF==0.14.2
future==0.16.0
googlemaps==2.5.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.14.2
oauthlib==2.0.7
passlib==1.7.1
pymongo==3.6.1
pyTelegramBotAPI==3.6.2
python-twitter==3.4.1
pytz==2018.3
requests==2.18.4
requests-oauthlib==0.8.0
requests-toolbelt==0.8.0
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1
WTForms==2.1
